\name{BigTree}
\alias{BigTree}
\alias{BigTreeDat}

\title{Example dataset for CAIC analyses}
\description{
BigTree is an artificially created dataset used to illustrate methods in the CAIC package.
}
\usage{
data(BigTree)
}
\format{
The dataset provides a phylogeny (\code{BigTree}) and a data frame (\code{BigTreeDat}). The phylogeny is a 200 tip tree containing some polytomies and the dataframe provides four continuous variables (yv, xc1, xc2, xc3) and one binary categorical variable (xf) for each of the tips in the tree. The dataset contains some NA values.

One reason this dataset  has been preserved as a dataset is that it also provides a reference dataset back to analyses performed in the original CAIC and MacroCAIC programs for Mac Classic. 
}


\seealso{ caic, macrocaic }

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{datasets}

