\name{profileLik.lambda}
\alias{profileLik.lambda}
\alias{profileLik.kappa}
\alias{lik.lambda}
\alias{lik.kappa}
\alias{maxLik.lambda}
\alias{maxLik.kappa}
\alias{lam.test.single}
\alias{order.V}
\alias{order.D}

%- Also NEED an '\alias' for EACH other topic documented here.
\title{Likelihood of branch length transformation parameters for single traits.}
\description{
The two parameters lambda and kappa are branch length transformations of use in phylogenetic GLM models. The two main functions (\code{lik.lambda} and \code{lik.kappa}) calculate the likelihood of lambda and kappa values, given a phylogeny and either a vector of data or a named variable in a data frame.  The \code{profileLik} functions calculate and plot a likelihood surface for a range of lambda and kappa values and the \code{maxLik} functions are a wrapper to \code{optimize} to calculate the maximum likelihood values of the parameters given  data for a a single trait and the phylogeny.

All of the likelihood functions expect a variance matrix with a matching vector of data for the tip names in the phylogeny and \code{prune.single} is a helper function to provide these inputs from a data frame and a variance covariance matrix. The function \code{lam.test.single} is a simple wrapper around \code{prune.single} and the likelihood functions that prints out a log likelihood ratio test of for the named variable of the maximum likelihood value of lambda against the extreme values of lambda =0 and lambda=1.

The two order functions are used by \code{prune.single} to order the data and variance covariance matrix alphabetically and will probably be concealed from users shortly.
}
\usage{
profileLik.lambda(x, V, data, at=seq(0, 1, by=0.01), plot=TRUE, ...) 
lik.lambda(x, V, data, lambda)
maxLik.lambda(x, V, data)

profileLik.kappa(x, V.arr, data, at=seq(0, 1, by=0.01), plot=TRUE, ...)
lik.kappa(x, V.arr, data, kappa)
maxLik.kappa(x, V.arr, data) 

lam.test.single(x, data, V, pretty=TRUE)

order.V(V)
order.D(data)
}

%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{A named vector of data values, where the names are used to match the values to the tips of the variance covariance matrix or a variable to look for in \code{data}.}
  \item{data}{An optional data frame, containing a variable identified using \code{x} to use for likelihood calculation. The row names must be set to contain the tip labels of the phylogeny}
  \item{V}{A variance-covariance matrix, although the array form used below is also accepted.}
  \item{V.arr}{A variance-covariance array holding individual edge lengths, as generated by \code{vcv.phylo.array}.}
  \item{lambda}{A value of lambda for which to calculate the likelihood given the data and phylogeny.}
  \item{kappa}{A value of kappa for which to calculate the likelihood given the data and phylogeny.}
  \item{at}{Values of kappa or lambda for which to calculate log likelihoods}
  \item{plot}{Logical indicating whether to plot the likelihood surface.}
  \item{pretty}{Currently unimplemented argument.}
  \item{...}{Further arguments to \code{plot}.}
}
\details{
Lambda is a proportional rescaling of the shared path lengths of the tips. Where \eqn{lambda = 0}, the internal path lengths are all zero and variance only accumulates along the paths to the tips - in this case, phylogenetic structure does not account for any of the variation in the data. Kappa is a power transformation of the individual shared edge lengths between tips and requires a array form of the VCV matrix in order to retain individual edge lengths rather than shared path lengths. 

[ stuff in here about upper limits...]

}
\value{
Both \code{lik.lambda} and \code{lik.kappa} return a list:
 \item{ll} Calculated log likelihood for the supplied lambda or kappa.
 \item{mu} Estimated phylogenitically weighted mean.
 \item{s2} Estimated variance of the mean.
The \code{profile} functions invisibly return a data frame of the \code{at} values and their associated log likelihoods. 
The \code{maxLik} functions return \code{optimize} output: namely, a list with components maximum and objective which give the location of the maximum likelihood and the value of the lambda or kappa at that point.

}
\references{ ~put references to the literature/web site here ~ }
\author{Rob Freckleton; David Orme; Gavin Thomas}
\seealso{\code{\link{pglm}}}
\examples{

data(shorebird)
shorebird.VCV <- vcv.phylo(shorebird.tree)
shorebird.VCVA <- vcv.phylo.array(shorebird.tree)

lik.lambda("M.Mass", shorebird.VCV, shorebird.data, 0.3)
M.Mass.lamSurf <- profileLik.lambda("M.Mass", shorebird.VCV, shorebird.data)
M.Mass.lamML <- maxLik.lambda("M.Mass", shorebird.VCV, shorebird.data)
abline(v=M.Mass.lamML)

lik.kappa("M.Mass", shorebird.VCVA, shorebird.data, 0.3)
M.Mass.kapSurf <- profileLik.kappa("M.Mass", shorebird.VCVA, shorebird.data)
M.Mass.kapML <- maxLik.kappa("M.Mass", shorebird.VCVA, shorebird.data)
abline(v=M.Mass.kapML)

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ utilities }
\keyword{ htest }% __ONLY ONE__ keyword per line
